Further Testing
I did further testing, and found that the airport list may not be as
informative as we may have already imagined. For example, Yaounde has
only one airport, which is the Nsimalen International Airport, but the
airport file has another airport which it labels Yaounde Airport.

Hopefully we find a nice way to solve this problem.